@@ -4,6 +4,7 @@ import com.example.todos.entities.User;
 import com.example.todos.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,8 +13,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.security.Principal;
-
 @Controller
 public class AuthController {
 
@@ -21,8 +20,9 @@ public class AuthController {
     private UserService userService;
 
     @RequestMapping(value = "/login", method = RequestMethod.GET)
-    public ModelAndView login(Principal principal) {
-        if (principal != null && ((Authentication) principal).isAuthenticated()) {
+    public ModelAndView login() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null && auth.isAuthenticated() && !"anonymousUser".equals(auth.getPrincipal())) {
             return new ModelAndView("redirect:/todos");
         }
         ModelAndView modelAndView = new ModelAndView();
@@ -33,8 +33,7 @@ public class AuthController {
     @GetMapping(value = "/registration")
     public ModelAndView registration() {
         ModelAndView modelAndView = new ModelAndView();
-        User user = new User();
-        modelAndView.addObject("user", user);
+        modelAndView.addObject("user", new User());
         modelAndView.setViewName("registration");
         return modelAndView;
     }
